lr2/requirements.txt
fastapi
uvicorn
pydantic
pyjwt
passlib
bcrypt==4.0.1
python-multipart
pytest
httpx
pyyaml